+++ tsup.config.ts
@@ -1,5 +1,0 @@
-import type { Options } from 'tsup'
-export const tsup: Options = {
-  entryPoints: ['src/index.ts'],
-  dts: true,
-}