+++ vite.config.ts
@@ -1,20 +1,25 @@
-import path from 'path'
+import { fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'
-import { name as PackageName } from './package.json'
+import dts from 'vite-plugin-dts'
+import packageJson from './package.json' with { type: 'json' }
 
-const camelCaseName = PackageName.replace(/-[a-zA-Z]/g, (m) =>
+const camelCaseName = packageJson.name.replace(/-[a-zA-Z]/g, (m) =>
   m.slice(1).toUpperCase()
 )
 const prod = process.env.NODE_ENV === 'production'
 export default defineConfig({
+  plugins: [dts()],
   build: {
     lib: {
-      entry: path.resolve(__dirname, 'src/index.ts'),
+      entry: fileURLToPath(new URL('./src/index.ts', import.meta.url)),
       name: camelCaseName,
       formats: ['es', 'cjs', 'umd'],
-      fileName: (format) => `${PackageName}.${format}${prod ? '.min' : ''}.js`,
+      fileName: (format) =>
+        format === 'cjs'
+          ? `${packageJson.name}.cjs`
+          : `${packageJson.name}.${format}.js`,
     },
     minify: prod,
-    emptyOutDir: prod,
+    emptyOutDir: true,
   },
 })